career_guide/requirements.txt
openai
django-environ